refactor(protocol): split the opportunity tools into registry, listing, and card helpers
